requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
bcrypt==3.1.6
blinker==1.4
cffi==1.12.2
Click==7.0
constantly==15.1.0
cryptography==2.6.1
cssselect==1.0.3
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gunicorn==19.9.0
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.10
lxml==4.3.3
MarkupSafe==1.1.1
parsel==1.5.1
Pillow==6.1.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
PyDispatcher==2.0.5
pyOpenSSL==19.0.0
queuelib==1.5.0
Scrapy==1.6.0
service-identity==18.1.0
six==1.12.0
SQLAlchemy==1.3.2
Twisted==17.1.0
w3lib==1.20.0
Werkzeug==0.15.2
WTForms==2.2.1
zope.interface==4.6.0
